requirements.txt
celery==4.3.0
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
idna==2.8
lxml==4.4.1
mysqlclient==1.4.4
requests==2.22.0
soupsieve==1.9.4
sqlparse==0.3.0
urllib3==1.25.6
virtualenv==16.7.5
djangorestframework==3.11.0
python-decouple==3.3
vine
